requirements.txt
# Required for serve.py
certifi==2019.11.28
Click==7.0
Flask==1.1.1
Flask-Limiter==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
limits==1.5
MarkupSafe==1.1.1
numpy==1.18.1
pbr==5.4.4
python-dateutil==2.8.1
six==1.13.0
Werkzeug==0.16.0
# Required if hosting with gunicorn
gunicorn==20.0.4
# Required to run other py scripts
beautifulsoup4==4.8.2
scikit-learn==0.22.1
scipy==1.4.1
soupsieve==1.9.5
tqdm==4.42.1